api/requirements.txt
flask==2.3.2
sqlalchemy==2.0.15
python-dotenv==1.0.0
gunicorn==20.1.0
flask-login==0.6.2
werkzeug==2.3.6
psycopg2-binary==2.9.6
